# packages/schemas/raw/openapi.yaml
...
openapi: 3.1.0
info:
  title: Relay Canonical Schemas
  version: 0.1.0
  description: |
    Canonical control-plane envelope schemas for the Relay agent reliability
    OS. Generated from this single OpenAPI 3.1 document into Pydantic v2
    (Python) and TS type aliases. The drift check (VAL-W1-035) keeps the
    generated trees byte-equal with the YAML on every commit.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /v1/state/transition:
    post:
      operationId: postV1StateTransition
      summary: Apply a state-engine transition
      description: |
        State-engine HTTP boundary. Validates the three-anchor
        handoff (scope_id, actor_identity_hash, manifest_commit_hash;
        keystone #4) BEFORE forwarding to compare_and_set_state.
        Stale or cross-project handoffs are rejected with the
        appropriate GUARD_FAILED code (e.g. RELAY-GATE-021).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Transition applied; canonical scope_state updated.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Malformed body, missing anchors, or invalid transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: |
            Actor kind not allowed for the requested transition
            (ACTOR_NOT_ALLOWED). Canonical ErrorEnvelope; the structured
            reason rides in details.reason.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: |
            Scope not found (UNKNOWN_SCOPE). Canonical ErrorEnvelope; the
            structured reason rides in details.reason.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: |
            Stale three-anchor handoff (RELAY-GATE-021) or state-machine
            conflict (EXPECTED_FROM_MISMATCH / TERMINAL_STATE). Canonical
            ErrorEnvelope; the structured reason rides in details.reason.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: |
            Transition rejected (INVALID_TRANSITION, GUARD_FAILED, or the
            fail-closed UNKNOWN_GUARD drift defense). Canonical
            ErrorEnvelope; the structured reason rides in details.reason.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # =========================================================================
  # V3 M2 F02 hosted-only 501 stubs (VAL-V3M2-004, VAL-V3M2-005)
  # =========================================================================
  # [OUT-OF-SCOPE-PRIVATE] These 5 routes belong to the private
  # ``relay-platform`` hosted control plane. The OSS sidecar exposes
  # them only as 501 stubs returning RELAY-HOSTED-ONLY with
  # blocked_surface=hosted_control_plane. Implementing the actual
  # hosted logic in OSS is a P0 boundary violation per CLAUDE.md
  # "Repository topology" and .ops boundaries.md DEFERRED item #3.
...
components:
  schemas:
...
    RelayErrorCodeStr:
      type: string
      pattern: '^RELAY-[A-Z]+-[0-9]{3}$'
      description: |
        Canonical Relay error-code wire form per VAL-W1-029. Known values
        enumerated in packages/schemas/raw/relay-error-codes.yaml.

    # --- W1.1 envelopes -------------------------------------------------------
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      description: |
        Canonical Relay error envelope. Spec B.4 lines 3392-3408.
      required:
        - schema_version
        - code
        - http_status
        - blocked_surface
        - retry_advice
        - request_id
        - trace_id
      properties:
        schema_version:
          type: string
          const: relay.error.v1
        code:
          $ref: '#/components/schemas/RelayErrorCodeStr'
        http_status:
          type: integer
          minimum: 400
          maximum: 599
        blocked_surface:
          type: string
          minLength: 1
        retry_advice:
          type: string
          enum:
            - do_not_retry
            - after_fix
            - after_retry_after
            - after_split
            - after_recapture
            - after_re_auth
        request_id:
          type: string
          minLength: 1
        trace_id:
          type: string
          minLength: 1
        message:
          type: [string, "null"]
        details:
          type: object
          additionalProperties: true
          default: {}

    # =========================================================================
    # v0.2 OSS completeness, M01 w1-1 (added 2026-05-16): canonical envelopes
    # for the 13 SQL tables in packages/schemas/sql/0004_v2_canonical_tables.sql.
    # =========================================================================
...